refactor(embeds): move embeds into their own package
