+++ packages/embeds/graphql/schema.js
@@ -1,0 +1,11 @@
+module.exports = `
+
+schema {
+  query: queries
+}
+
+type queries {
+  embed(id: ID!, embedType: EmbedType!): Embed!
+}
+
+`
