+++ packages/embeds/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  graphql: require('./graphql'),
+  ...require('./lib')
+}
