+++ servers/publikator/graphql/schema.js
@@ -16,7 +16,6 @@ type queries {
     orderBy: RepoOrderBy
   ): RepoConnection!
   repo(id: ID!): Repo!
-  embed(id: ID!, embedType: EmbedType!): Embed!
 
   """
   This query is a cached version of repos query. It uses cached information
